# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Swagger API Server"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "show posts",
                "description": "Display a listing of posts.",
                "operationId": "1e27aa05babd015976bc44487daaa8c1",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "List of posts",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostList"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create a post",
                "description": "Create a new post",
                "operationId": "b4a1533fe9fd61be03f965b9a94cde35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "I finally got a job"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "i am so glad"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "show post",
                "description": "Display a chosen post.",
                "operationId": "00b545a017fd7891b826b7b87346b1bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update a post",
                "description": "Update a post",
                "operationId": "0c5a81d0918cb826cd6b0651adf08e92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "I finally got a job"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "i am so glad"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete a post",
                "description": "Delete a post",
                "operationId": "2aa88de65366a7ede229dd1f782af766",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/posts/{id}/comment": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "comment a post",
                "description": "comment a post",
                "operationId": "155d91e8ce1a6252fc6f7df4cdf40b4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "i am so glad"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Comment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/posts/{id}/like": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "like a post",
                "description": "like a post",
                "operationId": "62a54ed6eef8b0e978f3e2a6a816f3b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Like"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/posts/{id}/dislike": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "dislike a post",
                "description": "dislike a post",
                "operationId": "cad9da13c541b5e17b1d91b093086f4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostList": {
                "required": [
                    "id",
                    "title",
                    "text",
                    "category",
                    "comments_count",
                    "likes_count"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "group foto"
                    },
                    "text": {
                        "type": "string",
                        "example": "head too heavy"
                    },
                    "category": {
                        "type": "string",
                        "example": "summer"
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Comment": {
                "required": [
                    "id",
                    "text",
                    "user_firstname",
                    "user_lastname",
                    "post_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "That`s awesome"
                    },
                    "user_firstname": {
                        "type": "string",
                        "example": "Walter"
                    },
                    "user_lastname": {
                        "type": "string",
                        "example": "White"
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Post": {
                "required": [
                    "title",
                    "text",
                    "category",
                    "comments_count",
                    "likes_count"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "group foto"
                    },
                    "text": {
                        "type": "string",
                        "example": "head too heavy"
                    },
                    "user_firstname": {
                        "type": "string",
                        "example": "Walter"
                    },
                    "user_lastname": {
                        "type": "string",
                        "example": "White"
                    },
                    "comments": {
                        "description": "List of comments",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    }
                },
                "type": "object"
            },
            "Like": {
                "required": [
                    "user_id",
                    "post"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post": {
                        "$ref": "#/components/schemas/PostList"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}